+++ spec/acceptance/sources_spec.rb
@@ -65,6 +65,38 @@ describe 'Chocolatey Source' do
     end
   end
 
+  context 'MODULES-5898 - Add Admin Only to an Existing Source' do
+
+    before(:all) do
+      backup_config
+    end
+
+    after(:all) do
+      reset_config
+    end
+
+    windows_agents.each do | agent |
+
+      it 'Should Apply the Manifest' do
+        chocolatey_src = <<-PP
+          chocolateysource {'chocolatey':
+            ensure   => present,
+            location => 'https://chocolatey.org/api/v2',
+            admin_only => true,
+          }
+        PP
+
+        execute_manifest_on(agent, chocolatey_src, :catch_failures => true)
+      end
+
+      it 'Should now set the source as visible only to administrators' do
+        on(agent, config_content_command) do |result|
+          assert_match(/true/, get_xml_value("//sources/source[@id='chocolatey']/@adminOnly", result.output).to_s, 'Admin Only did not match')
+        end
+      end
+    end
+  end
+
   context 'MODULES-3037 - Add Source With All Options' do
 
     before(:all) do
@@ -86,6 +118,7 @@ describe 'Chocolatey Source' do
             user         => 'bob',
             password     => 'yes',
             bypass_proxy => true,
+            admin_only   => true,
           }
         PP
 
@@ -100,6 +133,7 @@ describe 'Chocolatey Source' do
           assert_match(/.+/, get_xml_value("//sources/source[@id='test']/@password", result.output).to_s, 'Password was not saved')
           assert_match(/false/, get_xml_value("//sources/source[@id='test']/@disabled", result.output).to_s, 'Disabled did not match')
           assert_match(/true/, get_xml_value("//sources/source[@id='test']/@bypassProxy", result.output).to_s, 'Bypass Proxy did not match')
+          assert_match(/true/, get_xml_value("//sources/source[@id='test']/@adminOnly", result.output).to_s, 'Admin Only did not match')
         end
       end
     end
@@ -293,6 +327,55 @@ describe 'Chocolatey Source' do
       it 'Should verify results' do
         on(agent, config_content_command) do | result |
           assert_match(/false/, get_xml_value("//sources/source[@id='chocolatey']/@bypassProxy", result.output).to_s, 'Bypass Proxy change did not match')
+        end
+      end
+    end
+  end
+
+  context 'MODULES-5898 - Change Existing Admin Only Setting' do
+
+    before(:all) do
+      backup_config
+    end
+    
+    after(:all) do
+      reset_config
+    end
+
+    windows_agents.each do | agent |
+      it 'Should apply a manifest to set admin_only' do
+        chocolatey_src = <<-PP
+          chocolateysource {'chocolatey':
+            ensure   => present,
+            location => 'https://chocolatey.org/api/v2',
+            admin_only => true,
+          }
+        PP
+
+        execute_manifest_on(agent, chocolatey_src, :catch_failures => true)
+      end
+
+      it 'Should verify the setup was added' do
+        on(agent, config_content_command) do | result |
+          assert_match(/true/, get_xml_value("//sources/source[@id='chocolatey']/@adminOnly", result.output).to_s, 'Admin Only setup did not match')
+        end
+      end
+
+      it 'Should apply a manifest to set admin_only to false' do
+        chocolatey_src_change = <<-PP
+          chocolateysource {'chocolatey':
+            ensure   => present,
+            location => 'https://chocolatey.org/api/v2',
+            admin_only => false,
+          }
+        PP
+
+        execute_manifest_on(agent, chocolatey_src_change, :catch_failures => true)
+      end
+
+      it 'Should verify results' do
+        on(agent, config_content_command) do | result |
+          assert_match(/false/, get_xml_value("//sources/source[@id='chocolatey']/@adminOnly", result.output).to_s, 'Admin Only change did not match')
         end
       end
     end
@@ -666,6 +749,54 @@ describe 'Chocolatey Source' do
       it 'Should verify results' do
         on(agent, config_content_command) do | result |
           assert_match(/false/, get_xml_value("//sources/source[@id='chocolatey']/@bypassProxy", result.output).to_s, 'Bypass Proxy change did not match')
+        end
+      end
+    end
+  end
+
+  context 'MODULES-5898 Remove Admin Only from an Existing Source' do
+
+    before(:all) do
+      backup_config
+    end
+
+    after(:all) do
+      reset_config
+    end
+
+    windows_agents.each do | agent |
+      chocolatey_src = <<-PP
+        chocolateysource {'chocolatey':
+          ensure   => present,
+          location => 'https://chocolatey.org/api/v2',
+          admin_only => true,
+        }
+      PP
+
+      chocolatey_src_remove = <<-PP
+        chocolateysource {'chocolatey':
+          ensure   => present,
+          location => 'https://chocolatey.org/api/v2',
+        }
+      PP
+
+      it 'Should apply a manifest' do
+        execute_manifest_on(agent, chocolatey_src, :catch_failures => true)
+      end
+
+      it 'Should verify setup' do
+        on(agent, config_content_command) do | result |
+          assert_match(/true/, get_xml_value("//sources/source[@id='chocolatey']/@adminOnly", result.output).to_s, 'Admin Only did not match')
+        end
+      end
+
+      it 'Should apply remove manifest' do
+        execute_manifest_on(agent, chocolatey_src_remove, :catch_failures => true)
+      end
+
+      it 'Should verify results' do
+        on(agent, config_content_command) do | result |
+          assert_match(/false/, get_xml_value("//sources/source[@id='chocolatey']/@adminOnly", result.output).to_s, 'Admin Only change did not match')
         end
       end
     end